refactor: rename "params" to "batch"
The current name for the "params" parameter of the batch functions is confusing. The term "params" is reserved for the struct-encoded function parameters in Sablier V2 Core. By re-using it in the batch functions, we have inadvertently given it a double meaning, i.e. "params" also means an array of "params".

To clear up the confusion, this commit renames the "params" parameter to "batch" in all batch functions. Basically, a "batch" means an array of params now.